Implementation of the notification service and refactoring of the logic of certain files

+++ src/main/java/com/example/LMS/services/impl/BookServiceImpl.java
@@ -3,6 +3,7 @@ package com.example.LMS.services.impl;
 import com.example.LMS.dtos.BookRequestDto;
 import com.example.LMS.dtos.BookResponseDto;
 import com.example.LMS.exceptions.BookNotFoundException;
+import com.example.LMS.exceptions.BusinessLogicException;
 import com.example.LMS.mappers.BookMapper;
 import com.example.LMS.models.Book;
 import com.example.LMS.models.Gender;
@@ -32,7 +33,12 @@ public class BookServiceImpl implements BookService {
          //check if the gender exists
         Gender gender = genderRepository.findById(request.genderId()).orElseThrow(() -> new IllegalArgumentException("gender not found"));
 
+        if(bookRepository.existsByTitle(request.title())){
+            throw new BusinessLogicException("Book with title '"+request.title()+"' already exists", "BOOK_ALREADY_EXISTS");
+        }
+
         Book book = bookMapper.toEntity(request);
+        book.setAvailableCopies(request.numOfCopies());
         book.setGender(gender);
 
         //save the book in the database
@@ -56,6 +62,12 @@ public class BookServiceImpl implements BookService {
         //check if the book exists
         Book book = findBookById(id);
 
+        int borrowedBook = book.getNumOfCopies() - book.getAvailableCopies();
+
+        if(request.numOfCopies() < borrowedBook){
+            throw new BusinessLogicException("The number of copies can't be less than the number of copies currently on loan", "BAD_REQUEST");
+        }
+
         book.setNumOfCopies(request.numOfCopies());
 
         Book savedBook = bookRepository.save(book);
@@ -75,7 +87,11 @@ public class BookServiceImpl implements BookService {
     public void deleteBookById(Long id) {
 
         //check if the book exists
-        findBookById(id);
+        Book book = findBookById(id);
+
+        if(book.getAvailableCopies() != book.getNumOfCopies()){
+            throw new BusinessLogicException("This book can't be removed because there is one or more copies that have been borrowed","NOT-REMOVABLE_BOOK");
+        }
 
         bookRepository.deleteById(id);
